refactor: move useEstimations hook back into 'features/stake'
